feat: 🎉 Stable version 2.0 of the Freelance Portfolio Manager (#49)
* feat: :sparkles: The "client" model is created

In this model, data is received and a validation is carried out to ensure that the data provided is correct.

* feat: :sparkles: Customer Service is created

This file handles CRUD with MongoDB.

* feat: :sparkles: DB configuration is performed

The .env file is also created for the credentials issue, these must be replaced by a super admin at the end of the project development.

* feat: :sparkles: CLI ingration is performed

The .env file is configured, the connection to the DB is tested, and tests are performed by entering, editing, deleting, and listing DB clients.

* feat: :sparkles: The proposal.js model is created

The proposal.js model is created where the data is received and validation is performed to ensure that it is entered correctly.

* feat: :sparkles: proposalService.js is created

Within this file, the functions to create, list all, list by client and update the status of the proposals are created.

* feat: :sparkles: The proposal management menu is created and linked to index.js.

The creation, modification, deletion and listing of the proposals section are tested, and it is also correctly linked to index.js and its respective Menu.

* feat: :sparkles: Project.js is created where the data is received and the entered data is validated.

* feat: :sparkles: ProjectService.js is created

Within the file, the functions for creating, listing, updating, and deleting projects are created.

* feat: :sparkles: The project management section is completed.

Project management is generated and tested, index.js and the DB are linked, all functionalities are tested and the user story is finalized with all its acceptance criteria.

* feat: :sparkles: A model contract is created for the entry and validation of the data provided.

* feat: :sparkles: Create contratoService.js with the required functions

* fix: :bug: 🔹 HU03 – Editar cliente - 🔹 HU02 – Listar clientes - 🔹 HU01 – Crear cliente

Bug fixes and more JSON Schema validations are added.

* fix: :bug: 🔹 HU10 – Eliminar propuesta - 🔹 HU09 – Cambiar estado de propuesta - 🔹 HU08 – Listar propuestas por cliente - 🔹 HU07 – Listar TODAS las propuestas - 🔹 HU06 – Crear propuesta

Bug fixes are being worked on within these user stories, JSONschema validations are being added, and functionalities are being inserted.

* feat: :art: The "menu" folder is deleted and everything is moved to "utils"

* feat: :zap: 🔹 HU16 – Eliminar proyecto - 🔹 HU15 – Cambiar estado del proyecto - 🔹 HU14 – Listar proyectos por cliente - 🔹 HU13 – Listar todos los proyectos - 🔹 HU12 – Crear proyecto manualmente

Bugs are fixed, validation via JSON Schema is also added, and several functionalities are added based on user stories and their acceptance criteria.

* feat: :zap: 🔹 HU20 – Listar todos los contratos - 🔹 HU19 – Ver contrato por proyecto - 🔹 HU18 – Crear contrato para un proyecto

Errors are corrected, Json Schema validation is added and small functionalities are added. Work must be done on eliminating contracts and the management of this section would be finalized.

* feat: :sparkles: 🔹 HU22 – Crear entregable para proyecto

The user story of creating the deliverables for the client associated with a project is closed.

* feat: :sparkles: 🔹 HU23 – Listar entregables por proyecto

It is allowed to list the deliverables of a specific project without errors.

* feat: :construction: Working 🔹 HU24 – Cambiar estado del entregable

The acceptance criterion is missing. If the status changes to rejected, it must allow a reason to be included.

* feat: :sparkles: 🔹 HU25 – Eliminar entregable

The deliverable is eliminated by meeting the acceptance criteria of the user story.

* feat: :sparkles: 🔹 HU26 – Volver al menú principal

You successfully return to the main menu without errors.

* feat: :sparkles: 🔹 HU24 – Cambiar estado del entregable

This user story is concluded by adding that If the status changes to rejected, it should allow a reason to be included.

* feat: :construction: User story 🔹 HU27 – Registrar ingreso financiero -  is being worked on.

The money entry is created, but there is an error that does not validate whether the project has an active contract or not at the time of making the money entry. This detail needs to be resolved.

* feat: :sparkles: 🔹 HU27 – Registrar ingreso financiero

* feat: :construction: We are working on a bug fix when you add money to a project.

* fix: :bug: A bug in this user story is fixed: 🔹 HU27 – Registrar ingreso financiero

* feat: :sparkles: 🔹 HU28 – Registrar egreso financiero

This user story is carried out and it is verified that all the established acceptance criteria are met. Additionally, it is tested to avoid errors and transactions are used to reverse if any problem is found along the way.

* feat: :zap: 🔹 HU29 – Ver balance por cliente - 🔹 HU31 – Listar movimientos financieros

These user stories are finalized and tests are performed to verify their correct operation.

* docs: :memo: README.md is added

* feat: :sparkles: Improvements are added and errors are corrected

* feat: :sparkles: Improvements are added and errors are corrected